requirements.txt
amqp==5.1.1
argcomplete==2.0.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.1.0
Beaker==1.10.0
beautifulsoup4==4.11.0
billiard==3.6.4.0
blivet==3.4.4
blivet-gui==2.3.0
Brlapi==0.8.3
Brotli==1.0.9
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
chrome-gnome-shell==0.0.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.6
cryptography==36.0.0
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
decorator==4.0.11
Deprecated==1.2.13
distlib==0.3.4
distro==1.6.0
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-filter==21.1
django-timezone-field==5.0
djangorestframework==3.13.1
executing==0.8.2
fedora-third-party==0.10
filelock==3.7.1
filetype==1.0.5
fros==1.1
future==0.18.2
gpg==1.15.1
huepy==1.2.1
humanize==3.13.1
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
ipython==8.3.0
jedi==0.18.1
joblib==1.1.0
kombu==5.2.4
langtable==0.0.58
libcomps==0.1.18
lxml==4.7.1
Mako==1.1.4
Markdown==3.3.7
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
mock==4.0.3
moviepy==1.0.3
mutagen==1.45.1
nftables==0.1
nltk==3.7
numpy==1.22.4
olefile==0.46
packaging==21.3
parso==0.8.2
Paste==3.5.0
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
productmd==1.33
proglog==0.1.10
prompt-toolkit==3.0.24
psutil==5.8.0
psycopg2-binary==2.9.3
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.4
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.14.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
Pygments==2.11.2
PyGObject==3.42.1
pykickstart==3.36
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
PySocks==1.7.1
pytest==7.1.2
python-augeas==1.1.0
python-crontab==2.6.0
python-dateutil==2.8.1
python-meh==0.50
python-xlib==0.31
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
redis==4.3.3
regex==2022.4.24
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==0.9.1
responses==0.21.0
rpm==4.17.0
rsa==4.8
schedule==1.1.0
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.3
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
systemd-python==234
telebot==0.0.4
Telethon==1.24.0
Tempita==0.5.2
terminator==2.1.1
textblob==0.17.1
tomli==2.0.1
tqdm==4.64.0
traitlets==5.1.1
tzdata==2022.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
websockets==10.1
wrapt==1.14.1
yt-dlp==2022.5.18
